refactor(server): align handler route module names
