+++ nodelite-server/src/handlers/install_routes.rs
