+++ nodelite-server/src/handlers/metrics_routes.rs
